refactor(split): move ble peripheral run to nrf mod
