+++ rmk/src/split/nrf/peripheral.rs
@@ -1,10 +1,16 @@
+use crate::ble::nrf::softdevice_task;
+use crate::split::driver::{SplitDriverError, SplitReader, SplitWriter};
+use crate::split::peripheral::SplitPeripheral;
+use crate::split::{SplitMessage, SPLIT_MESSAGE_MAX_SIZE};
+use crate::MatrixTrait;
+use core::mem;
 use defmt::error;
+use embassy_executor::Spawner;
+use nrf_softdevice::ble::gatt_server::set_sys_attrs;
+use nrf_softdevice::ble::peripheral::{advertise_connectable, ConnectableAdvertisement};
 use nrf_softdevice::ble::{gatt_server, Connection};
-
-use crate::split::{
-    driver::{SplitDriverError, SplitReader, SplitWriter},
-    SplitMessage, SPLIT_MESSAGE_MAX_SIZE,
-};
+use nrf_softdevice::ble::{set_address, Address, AddressType};
+use nrf_softdevice::{raw, Config, Softdevice};
 
 /// Gatt service used in split peripheral to send split message to central
 #[nrf_softdevice::gatt_service(uuid = "4dd5fbaa-18e5-4b07-bf0a-353698659946")]
@@ -44,6 +50,10 @@ impl<'a> SplitReader for BleSplitPeripheralDriver<'a> {
                 error!("BLE read error: {:?}", e);
                 SplitDriverError::BleError(1)
             })?;
+        defmt::info!(
+            "Got read from central in BleSplitPeripheralDriver read: {:?}",
+            message
+        );
         let message: SplitMessage = postcard::from_bytes(&message).map_err(|e| {
             error!("Postcard deserialize split message error: {}", e);
             SplitDriverError::DeserializeError
@@ -69,5 +79,136 @@ impl<'a> SplitWriter for BleSplitPeripheralDriver<'a> {
             SplitDriverError::BleError(1)
         })?;
         Ok(bytes.len())
+    }
+}
+
+/// Initialize and run the nRF peripheral keyboard service via BLE.
+///
+/// # Arguments
+///
+/// * `input_pins` - input gpio pins
+/// * `output_pins` - output gpio pins
+/// * `spawner` - embassy task spawner, used to spawn nrf_softdevice background task
+pub(crate) async fn initialize_nrf_ble_split_peripheral_and_run<
+    M: MatrixTrait,
+    const ROW: usize,
+    const COL: usize,
+>(
+    mut matrix: M,
+    central_addr: [u8; 6],
+    peripheral_addr: [u8; 6],
+    spawner: Spawner,
+) -> ! {
+    use defmt::info;
+    use embassy_futures::select::select3;
+    use nrf_softdevice::ble::gatt_server;
+
+    use crate::{
+        split::nrf::peripheral::{
+            BleSplitPeripheralDriver, BleSplitPeripheralServer, BleSplitPeripheralServerEvent,
+            SplitBleServiceEvent,
+        },
+        CONNECTION_STATE,
+    };
+
+    let ble_config = Config {
+        clock: Some(raw::nrf_clock_lf_cfg_t {
+            source: raw::NRF_CLOCK_LF_SRC_RC as u8,
+            rc_ctiv: 16,
+            rc_temp_ctiv: 2,
+            accuracy: raw::NRF_CLOCK_LF_ACCURACY_500_PPM as u8,
+            // External osc
+            // source: raw::NRF_CLOCK_LF_SRC_XTAL as u8,
+            // rc_ctiv: 0,
+            // rc_temp_ctiv: 0,
+            // accuracy: raw::NRF_CLOCK_LF_ACCURACY_20_PPM as u8,
+        }),
+        conn_gap: Some(raw::ble_gap_conn_cfg_t {
+            conn_count: 6,
+            event_length: 24,
+        }),
+        conn_gatt: Some(raw::ble_gatt_conn_cfg_t { att_mtu: 256 }),
+        gatts_attr_tab_size: Some(raw::ble_gatts_cfg_attr_tab_size_t {
+            attr_tab_size: raw::BLE_GATTS_ATTR_TAB_SIZE_DEFAULT,
+        }),
+        gap_role_count: Some(raw::ble_gap_cfg_role_count_t {
+            adv_set_count: 1,
+            periph_role_count: 4,
+            central_role_count: 4,
+            central_sec_count: 4,
+            _bitfield_1: raw::ble_gap_cfg_role_count_t::new_bitfield_1(0),
+        }),
+        gap_device_name: Some(raw::ble_gap_cfg_device_name_t {
+            p_value: "rmk_peripheral_board".as_ptr() as _,
+            current_len: "rmk_peripheral_board".len() as u16,
+            max_len: "rmk_peripheral_board".len() as u16,
+            write_perm: unsafe { mem::zeroed() },
+            _bitfield_1: raw::ble_gap_cfg_device_name_t::new_bitfield_1(
+                raw::BLE_GATTS_VLOC_STACK as u8,
+            ),
+        }),
+        ..Default::default()
+    };
+
+    let sd = Softdevice::enable(&ble_config);
+    set_address(
+        sd,
+        &Address::new(AddressType::RandomStatic, peripheral_addr),
+    );
+
+    {
+        // Use the immutable ref of `Softdevice` to run the softdevice_task
+        // The mumtable ref is used for configuring Flash and BleServer
+        let sdv = unsafe { nrf_softdevice::Softdevice::steal() };
+        defmt::unwrap!(spawner.spawn(softdevice_task(sdv)))
+    };
+
+    let server = defmt::unwrap!(BleSplitPeripheralServer::new(sd));
+
+    loop {
+        CONNECTION_STATE.store(false, core::sync::atomic::Ordering::Release);
+        let advertisement: ConnectableAdvertisement<'_> =
+            ConnectableAdvertisement::NonscannableDirected {
+                peer: Address::new(AddressType::RandomStatic, central_addr),
+            };
+        let conn = match advertise_connectable(sd, advertisement, &Default::default()).await {
+            Ok(conn) => conn,
+            Err(e) => {
+                defmt::error!("Split peripheral advertise error: {}", e);
+                continue;
+            }
+        };
+
+        // Set sys attr of peripheral
+        set_sys_attrs(&conn, None).ok();
+
+        let server_fut = gatt_server::run(&conn, &server, |event| match event {
+            BleSplitPeripheralServerEvent::Service(split_event) => match split_event {
+                SplitBleServiceEvent::MessageToCentralCccdWrite { notifications } => {
+                    info!("Split value CCCD updated: {}", notifications)
+                }
+                SplitBleServiceEvent::MessageToPeripheralWrite(message) => {
+                    match postcard::from_bytes::<SplitMessage>(&message) {
+                        Ok(message) => {
+                            info!("Message from central: {:?}", message);
+                            match message {
+                                SplitMessage::CentralState(state) => {
+                                    CONNECTION_STATE
+                                        .store(state, core::sync::atomic::Ordering::Release);
+                                    info!("Received server state");
+                                }
+                                _ => (),
+                            }
+                        }
+                        Err(e) => defmt::error!("Postcard deserialize split message error: {}", e),
+                    }
+                }
+            },
+        });
+
+        let mut peripheral = SplitPeripheral::new(BleSplitPeripheralDriver::new(&server, &conn));
+        let peripheral_fut = peripheral.run();
+        let matrix_fut = matrix.run();
+        select3(matrix_fut, server_fut, peripheral_fut).await;
     }
 }
